Adds TEMPO prognostic cloud fraction variables to the core_atmosphere Registry.
Prognostic cloud fraction (qaprog) and its tendendies are controled by the package
tempo_cldfrac_in, which is activated by setting the configuration flag
config_tempo_cldfrac = True when using TEMPO microphysics. If this flag is set to
True when using TEMPO, the cloudiness scheme is automatically set to cld_tempo,
which handles the prognostic cloud fraction.